% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square_val.R
\name{square_val}
\alias{square_val}
\title{Carré d'une valeur}
\usage{
square_val(x)
}
\arguments{
\item{x}{Valeur numérique}
}
\value{
un nombre
}
\description{
Carré d'une valeur
}
\examples{
square_val(2)
}
